# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title; generated import paths depend on it.
  title: Api
  version: 0.2.0
  description: LootBot multi-tenant store management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores owned by the current account
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Owned stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a store
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /stores/{storeId}:
    parameters:
      - $ref: "#/components/parameters/StoreId"
    get:
      operationId: getStore
      tags: [stores]
      summary: Read an owned store
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update an owned store
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Soft-delete an owned store and stop its bot
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Store archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /stores/{storeId}/bot:
    parameters:
      - $ref: "#/components/parameters/StoreId"
    get:
      operationId: getStoreBot
      tags: [telegram]
      summary: Read bot connection status without its token
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: connectStoreBot
      tags: [telegram]
      summary: Validate and connect a Telegram bot
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotTokenInput"
      responses:
        "200":
          description: Bot connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          $ref: "#/components/responses/UpstreamError"
    delete:
      operationId: disconnectStoreBot
      tags: [telegram]
      summary: Disconnect and stop a store bot
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Bot disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get real store counts and recent sales summary
      security:
        - sessionCookie: []
      parameters:
        - name: storeId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Store summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent store activity
      security:
        - sessionCookie: []
      parameters:
        - name: storeId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 25
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lootbot_session
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Sign in required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflicts with existing data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UpstreamError:
      description: Telegram could not be reached or rejected the token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    StoreInput:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        currency:
          type: string
          minLength: 3
          maxLength: 3
      required: [name, currency]
    StoreUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        currency:
          type: string
          minLength: 3
          maxLength: 3
    Store:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        currency:
          type: string
        botStatus:
          type: string
          enum: [disconnected, connected, error]
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, name, slug, currency, botStatus, createdAt]
...
    BotTokenInput:
      type: object
      additionalProperties: false
      properties:
        token:
          type: string
          minLength: 20
          maxLength: 256
      required: [token]
    BotStatus:
      type: object
      properties:
        connected:
          type: boolean
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        status:
          type: string
          enum: [disconnected, connected, error]
        lastError:
          type: ["string", "null"]
      required: [connected, username, firstName, status, lastError]
    DashboardSummary:
      type: object
      properties:
        storeId:
          type: string
          format: uuid
        productCount:
          type: integer
        categoryCount:
          type: integer
        publishedProductCount:
          type: integer
        orderCount:
          type: integer
        revenue:
          type: number
        botStatus:
          type: string
          enum: [disconnected, connected, error]
      required: [storeId, productCount, categoryCount, publishedProductCount, orderCount, revenue, botStatus]
    ActivityEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, action, summary, createdAt]