# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LootBot multi-tenant store management API
 * OpenAPI spec version: 0.2.0
 */
import type { DashboardSummaryBotStatus } from './dashboardSummaryBotStatus';

export interface DashboardSummary {
  storeId: string;
  productCount: number;
  categoryCount: number;
  publishedProductCount: number;
  orderCount: number;
  revenue: number;
  botStatus: DashboardSummaryBotStatus;
}